requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
croniter==6.0.0
cryptography==44.0.2
cssselect2==0.8.0
debtcollector==3.0.0
decorator==5.2.1
dnspython==2.7.0
dogpile.cache==1.3.4
eval_type_backport==0.2.2
eventlet==0.39.1
fasteners==0.19
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.6.0
futurist==3.1.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-vision==3.10.1
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
iso8601==2.1.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keystoneauth1==5.10.0
keystonemiddleware==10.9.0
kombu==5.5.3
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistral==20.0.0
mistral-lib==3.3.1
mistralai==1.7.0
msgpack==1.1.0
multidict==6.4.3
narwhals==1.35.0
netaddr==1.3.0
networkx==3.4.2
numpy==2.2.4
os-service-types==1.7.0
oslo.cache==3.10.1
oslo.concurrency==7.1.0
oslo.config==9.7.1
oslo.context==5.7.1
oslo.db==17.2.1
oslo.i18n==6.5.1
oslo.log==7.1.0
oslo.messaging==16.1.0
oslo.metrics==0.11.0
oslo.middleware==6.4.0
oslo.policy==4.5.1
oslo.serialization==5.7.0
oslo.service==4.1.1
oslo.utils==8.2.0
osprofiler==4.2.0
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
Paste==3.10.1
PasteDeploy==3.1.0
pathlib==1.0.1
pbr==6.1.1
pecan==1.6.0
pillow==11.1.0
ply==3.11
prettytable==3.16.0
prometheus_client==0.21.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycadf==4.0.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
pydyf==0.11.0
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.3
PyPDF2==3.0.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-keystoneclient==5.6.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
repoze.lru==0.7
requests==2.32.3
rfc3986==2.0.0
rich==13.9.4
Routes==2.5.1
rpds-py==0.24.0
rsa==4.9.1
setuptools==79.0.1
shellingham==1.5.4
simplegeneric==0.8.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
statsd==4.0.1
stevedore==5.4.1
streamlit==1.44.1
streamlit-custom-notification-box==0.1.2
tabulate==0.9.0
tenacity==9.1.2
testresources==2.0.2
testscenarios==0.5.0
testtools==2.7.2
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
together==1.5.5
toml==0.10.2
tooz==6.3.0
tornado==6.4.2
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
WebOb==1.8.9
wrapt==1.17.2
WSME==0.12.1
yappi==1.6.10
yaql==3.0.0
yarl==1.20.0
zipp==3.21.0
zopfli==0.2.3.post1
